siia-backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg[binary]==3.2.13
django-environ==0.11.2
ldap3==2.9.1
django-auth-ldap==4.8.0
django-cors-headers==4.4.0
